Fix cluster scoped CRDs and conditions
Fix a bug where cluster scoped CRDs could not be loaded, when a namespace
was selected. Now we are checking explicit if the CRD is cluster scoped
and ignoring the namespace when this is the case.

Fix a bug where kubenav crashes, when for the conditions of a CRD an
object was used instead of an array. Now we are checking if the
conditions field is an array or an object.
